refactor(harness-driver): declare the broker agent payload once
`GET /api/spawned` and the `agents` array of `GET /api/status` are the
same broker payload — both are serialized from one `WorkerRegistry::list`
call — but TypeScript declared them twice, and the copies had drifted:
`BrokerStatus.agents` was missing `sessionId`, `runtime_kind`, the
`native_harness_*` fields, and now `pending_messages`, so `getStatus()`
consumers could not reach them without an unsafe cast.

Move `ListAgent` to the wire-contract module and point
`BrokerStatus.agents` at it, so the two surfaces cannot drift again.
`types.ts` re-exports `ListAgent`, keeping existing import paths valid.
